// frontend/src/api/generated/serializers/User.ts
// Typelizer digest 494de170e8b658bf5e08c3e8855ec1d4
//
// DO NOT MODIFY: This file was automatically generated by Typelizer.

type User = {
  id: string;
  name: string;
  email: string;
  role: "user" | "superadmin";
  locale: string;
  confirmedAt: string | null;
  insertedAt: string;
  hasPassword: boolean;
}

export type { User };